reentrega

+++ codigo.js
@@ -1,119 +1,70 @@
-const ANUAL= 365
+const ANUAL = 365
 
+let opciones;
 
-const OPCION1 = {
-
-
-    montoMaximo: 1000000,
-    plazo: 30,
-    tasa: 0.90
-
-
-}
-const OPCION2 = {
-
-
-    montoMaximo: 2000000,
-    plazo: 60,
-    tasa: 0.95
-
-
-}
-const OPCION3 = {
-
-
-    montoMaximo: 3000000,
-    plazo: 90,
-    tasa: 0.99
-
+function cargarOpciones() {
+    fetch("./opciones.json")
+        .then((response) => response.json())
+        .then((json) => opciones = json);
 
 }
 
-
-const OPCIONES = [OPCION1, OPCION2, OPCION3];
-
-
-function calcularOpcion (montoDeInversion) {
-    return OPCIONES.find ((opcion) => opcion.montoMaximo >=montoDeInversion ) 
-    
-
-
+function calcularOpcion(montoDeInversion) {
+    return opciones.find((opcion) => opcion.montoMaximo >= montoDeInversion)
 }
 
-
-function calculoFinal (montoDeInversion, opcion){
-    return (((montoDeInversion * opcion.tasa) / ANUAL)* opcion.plazo);
-    
-
-
+function calculoFinal(montoDeInversion, opcion) {
+    return (((montoDeInversion * opcion.tasa) / ANUAL) * opcion.plazo);
 }
 
+function validarEdad() {
+    const edadUsuario = document.getElementById("edad").value;
+    const guardaEdad = {
+        edad: edadUsuario,
+    }
+    // console.log(edadUsuario >= 18 ? "podés ingresar" : "No podés Ingresar, sos menor");
 
+    if (edadUsuario < 18) {
+        Swal.fire({
+            title: 'Eres Menor!',
+            text: "No puedes Continuar",
+            icon: 'error',
+            confirmButtonText: 'Aceptar',
+        });
+        return false;
+    } else {
+        return true;
+    }
+}
 
-function calcular (event){
-    event.preventDefault();        
+function calcular(event) {
+    event.preventDefault();
 
+    if (validarEdad() == false) {
+        return;
+    }
 
     const nombreUsuario = document.getElementById("nombre").value;
-    
+
     const guardaNombre = {
         nombre: nombreUsuario,
     }
 
-
     const usuarioJSON = JSON.stringify(guardaNombre);
     localStorage.setItem("nombreUsuario", usuarioJSON);
-
 
     const nombreUsuarioEnLocalStorage = localStorage.getItem("nombreUsuario");
     const usuarioObjeto = JSON.parse(nombreUsuarioEnLocalStorage);
     const nombreLS = document.getElementById("nombre").value;
 
-
-    
-
-
     const montoDeInversion = document.getElementById("montoDeInversion").value;
 
+    const opcionElegida = calcularOpcion(montoDeInversion)
 
-    const opcionElegida = calcularOpcion (montoDeInversion)
-
-
-    const resultado = calculoFinal (montoDeInversion,opcionElegida);
+    const resultado = calculoFinal(montoDeInversion, opcionElegida);
     const textoResultado = document.getElementById("textoResultado");
     textoResultado.textContent = nombreUsuario + "  El interes por " + opcionElegida.plazo + "   dias es de   " + resultado;
 }
-const boton= document.getElementById("boton");
-boton.addEventListener("click", calcular);
-
-
-    
-const sweetAlert = document.querySelector("#sweetAlert");
-
-sweetAlert.addEventListener("click", (event) => {
-    event.preventDefault();
-    const edadUsuario = document.getElementById("edad").value;
-const guardaEdad = {
-    edad: edadUsuario,
-}
-
-console.log (edadUsuario >= 18 ? "podés ingresar" : "No podés Ingresar, sos menor");
-    
-    
-    if (edadUsuario < 18){
-
-        
-    Swal.fire({
-        title: 'Eres Menor!',
-        text: "No puedes Continuar",
-        icon: 'error',
-        confirmButtonText: 'Aceptar',
-        
-        
-    })
-
-    }
-})
 
 async function mostrarClima() {
     const url = 'https://weatherapi-com.p.rapidapi.com/current.json?q=Argentina'
@@ -124,18 +75,20 @@ async function mostrarClima() {
             'X-RapidAPI-Host': 'weatherapi-com.p.rapidapi.com'
         }
     };
-    
+
     try {
         const response = await fetch(url, options);
         const result = await response.json();
-        const clima= result.current.temp_c;
+        const clima = result.current.temp_c;
         const climatologia = document.getElementById("climatologia");
         climatologia.textContent = clima + "  Grados"
-
-        console.log(result);
     } catch (error) {
         console.error(error);
     }
 }
 
+cargarOpciones();
 mostrarClima();
+
+const boton = document.getElementById("boton");
+boton.addEventListener("click", calcular);
